requirements-api.txt
opencv-python
numpy
torch
torchvision
fastapi
uvicorn
python-multipart
Pillow
transformers
requests
python-dotenv
pygame
git+https://github.com/huggingface/transformers.git
datasets
accelerate
https://pypi.org/simple/
bitsandbytes
peft